+++ decidim-admin/app/assets/config/decidim_admin_manifest.js
@@ -2,4 +2,5 @@
 //= link decidim/admin/managed_users.js
 //= link decidim/admin/resources_permissions.js
 //= link decidim/admin/welcome_notification.js
+//= link decidim/admin/newsletters.js
 //= link decidim/admin/application.css