+++ decidim-comments/app/models/decidim/comments/comment.rb
@@ -98,6 +98,13 @@ module Decidim
         Decidim::Comments::CommentSerializer
       end
 
+      def self.newsletter_participant_ids(space)
+        Decidim::Comments::Comment.includes(:root_commentable).not_hidden
+                                  .where("decidim_comments_comments.decidim_author_id IN (?)", Decidim::User.where(organization: space.organization).pluck(:id))
+                                  .where("decidim_comments_comments.decidim_author_type IN (?)", "Decidim::UserBaseEntity")
+                                  .map(&:author).pluck(:id).flatten.compact.uniq
+      end
+
       private
 
       # Private: Check if commentable can have comments and if not adds