+++ decidim-core/db/migrate/20190325145349_add_extended_data_to_newsletters.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddExtendedDataToNewsletters < ActiveRecord::Migration[5.2]
+  def change
+    add_column :decidim_newsletters, :extended_data, :jsonb, default: {}
+  end
+end